+++ src/components/AdminSettings.vue
@@ -2,8 +2,8 @@
 	<div id="openproject_prefs" class="section">
 		<TermsOfServiceUnsigned :is-all-terms-of-service-signed-for-user-open-project="isAllTermsOfServiceSignedForUserOpenProject" />
 		<SettingsTitle is-setting="admin" />
-		<NcNoteCard v-if="!isAdminAuditConfigurationSetUpCorrectly" class="audit-info-card" type="info">
-			<p class="audit-info-card--info" v-html="getAdminAuditConfigurationHint" /> <!-- eslint-disable-line vue/no-v-html -->
+		<NcNoteCard v-if="!isAdminAuditConfigurationSetUpCorrectly" class="note-card" type="info">
+			<p class="note-card--info-description" v-html="getAdminAuditConfigurationHint" /> <!-- eslint-disable-line vue/no-v-html -->
 		</NcNoteCard>
 		<div class="openproject-server-host">
 			<FormHeading index="1"
@@ -225,10 +225,12 @@
 								t('integration_openproject', 'The app will never delete files or folders, even if you deactivate this later.')
 							}}
 						</p>
-						<ProjectFolderError
-							v-if="projectFolderSetupError !== null"
-							:project-folder-set-up-error-message-description="projectFolderSetUpErrorMessageDescription(projectFolderSetupError)"
-							:project-folder-set-up-error="projectFolderSetupError" />
+						<NcNoteCard v-if="projectFolderSetupError !== null" class="note-card" type="error">
+							<p class="note-card--title">
+								<b>{{ projectFolderSetupError }}</b>
+							</p>
+							<p class="note-card--error-description" v-html="projectFolderSetUpErrorMessageDescription(projectFolderSetupError)" /> <!-- eslint-disable-line vue/no-v-html -->
+						</NcNoteCard>
 						<div class="form-actions">
 							<NcButton v-if="projectFolderSetupError === null"
 								type="primary"
@@ -257,10 +259,12 @@
 					<div class="project-folder-status-value">
 						<b>{{ t('integration_openproject','Automatically managed folders:') }}</b> {{ opUserAppPassword ? t('integration_openproject', 'Active') : t('integration_openproject', 'Inactive') }}
 					</div>
-					<ProjectFolderError
-						v-if="state.app_password_set && !isProjectFolderSetupCorrect"
-						:project-folder-set-up-error-message-description="projectFolderSetUpErrorMessageDescription(state.project_folder_info.errorMessage)"
-						:project-folder-set-up-error="state.project_folder_info.errorMessage" />
+					<NcNoteCard v-if="state.app_password_set && !isProjectFolderSetupCorrect" class="note-card" type="error">
+						<p class="note-card--title">
+							<b>{{ state.project_folder_info.errorMessage }}</b>
+						</p>
+						<p class="note-card--error-description" v-html="projectFolderSetUpErrorMessageDescription(state.project_folder_info.errorMessage)" /> <!-- eslint-disable-line vue/no-v-html -->
+					</NcNoteCard>
 					<div class="form-actions">
 						<NcButton
 							data-test-id="edit-project-folder-setup"
@@ -370,7 +374,6 @@ import FormHeading from './admin/FormHeading.vue'
 import CheckBox from '../components/settings/CheckBox.vue'
 import SettingsTitle from '../components/settings/SettingsTitle.vue'
 import { F_MODES, FORM, USER_SETTINGS } from '../utils.js'
-import ProjectFolderError from './admin/ProjectFolderError.vue'
 import TermsOfServiceUnsigned from './admin/TermsOfServiceUnsigned.vue'
 import dompurify from 'dompurify'
 export default {
@@ -388,7 +391,6 @@ export default {
 		RestoreIcon,
 		CheckBox,
 		NcCheckboxRadioSwitch,
-		ProjectFolderError,
 		TermsOfServiceUnsigned,
 		NcNoteCard,
 	},
@@ -1194,9 +1196,9 @@ export default {
 			padding-left: 5px;
 		}
 	}
-	.audit-info-card {
+	.note-card {
 		max-width: 900px;
-		&--info {
+		&--info-description, &--error-description {
 			.link {
 				color: #1a67a3 !important;
 				font-style: normal;